requirements.txt
requests
beautifulsoup4
markdownify
python-slugify
tqdm
python-dotenv
openai
tiktoken
python-dotenv
